Minor optimizations

+++ api/src/lib.rs
@@ -4,7 +4,7 @@ use shopify_function_wasm_api_core::{NanBox, ValueRef};
 extern "C" {
     // Read API.
     fn shopify_function_input_get() -> u64;
-    fn shopify_function_input_get_val_len(scope: u64) -> u32;
+    fn shopify_function_input_get_val_len(scope: u64) -> usize;
     fn shopify_function_input_read_utf8_str(src: usize, out: *mut u8, len: usize);
     fn shopify_function_input_get_obj_prop(scope: u64, ptr: *const u8, len: usize) -> u64;
     fn shopify_function_input_get_at_index(scope: u64, index: u32) -> u64;
